+++ EnchantmentsEnhance-Core/src/main/kotlin/org/pixeltime/enchantmentsenhance/event/enchantment/shear/Shearer.kt
@@ -16,13 +16,15 @@
  *
  */
 
-package org.pixeltime.enchantmentsenhance.event.enchantment.gear
+package org.pixeltime.enchantmentsenhance.event.enchantment.shear
 
+import org.bukkit.entity.Player
 import org.bukkit.entity.Sheep
 import org.bukkit.event.EventHandler
 import org.bukkit.event.EventPriority
 import org.bukkit.event.block.Action
 import org.bukkit.event.player.PlayerInteractEvent
+import org.bukkit.event.player.PlayerShearEntityEvent
 import org.bukkit.inventory.ItemStack
 import org.pixeltime.enchantmentsenhance.listener.EnchantmentListener
 import org.pixeltime.enchantmentsenhance.manager.SettingsManager
@@ -40,18 +42,32 @@ class Shearer : EnchantmentListener() {
     @EventHandler(ignoreCancelled = true, priority = EventPriority.MONITOR)
     fun onInteract(playerInteractEvent: PlayerInteractEvent) {
         val player = playerInteractEvent.player
-
-        if (playerInteractEvent.action == Action.LEFT_CLICK_AIR) {
+        if ((playerInteractEvent.action == Action.RIGHT_CLICK_BLOCK) || (playerInteractEvent == Action.RIGHT_CLICK_AIR)) {
             val level = getLevel(player)
             if (level > 0) {
-                val int1 = SettingsManager.enchant.getInt("shearer.$level.radius")
-                for (entity in player.getNearbyEntities(int1.toDouble(), int1.toDouble(), int1.toDouble())) {
-                    if (entity is Sheep) {
-                        if (!entity.isSheared) {
-                            entity.isSheared = true
-                            entity.world.dropItem(entity.location, ItemStack(XMaterial.WHITE_WOOL.parseMaterial(), 1, entity.color.woolData.toShort()))
-                        }
-                    }
+                shear(player, level)
+            }
+        }
+    }
+
+    @EventHandler
+    fun playerShearEvent(event: PlayerShearEntityEvent) {
+        if (event.entity is Sheep) {
+            val player = event.player
+            val level = getLevel(player)
+            if (level > 0) {
+                shear(player, level)
+            }
+        }
+    }
+
+    fun shear(player: Player, level: Int) {
+        val int1 = SettingsManager.enchant.getInt("shearer.$level.radius")
+        for (entity in player.getNearbyEntities(int1.toDouble(), int1.toDouble(), int1.toDouble())) {
+            if (entity is Sheep) {
+                if (!entity.isSheared) {
+                    entity.isSheared = true
+                    entity.world.dropItem(entity.location, ItemStack(XMaterial.WHITE_WOOL.parseMaterial(), 1, entity.color.woolData.toShort()))
                 }
             }
         }